Move prefix range filter definitions from old files to new ones and fix includes
